Make rust-analyzer work with workspaces
