+++ workspaces/icecap-host/src/icecap-host/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+authors = ["The Veracruz Development Team"]
+description = "A dummy package"
+edition = "2018"
+name = "linux-host"
+version = "0.3.0"
+
+[dependencies]
+policy-utils = { path = "../../crates/policy-utils", features = [ "icecap" ] }
+tests = { path = "../../crates/tests", features = [ "icecap" ] }
+veracruz-client = { path = "../../crates/veracruz-client", features = [ "icecap", "cli" ] }
+veracruz-server = { path = "../../crates/veracruz-server", features = [ "icecap", "cli" ] }
