+++ workspaces/icecap-runtime/src/icecap-runtime/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+authors = ["The Veracruz Development Team"]
+description = "The trusted Veracruz runtime's command-and-control module.  Co-ordinates and directs the rest of the runtime."
+edition = "2018"
+name = "runtime-manager-enclave"
+version = "0.3.0"
+resolver = "2"
+
+[features]
+default = [ "icecap" ]
+icecap = [ "runtime-manager/icecap" ]
+icecap-lkvm = []
+icecap-qemu = []
+
+[dependencies]
+runtime-manager = { path = "../../crates/runtime-manager" }
