+++ workspaces/nitro-host/.cargo/config.toml
@@ -1,0 +1,8 @@
+[build]
+rustflags = [
+    "-C", "link-args=-Wl,--build-id=none"
+]
+
+[env]
+CC_x86_64_unknown_linux_gnu = "gcc"
+CC_aarch64_unknown_linux_gnu = "gcc"
