+++ workspaces/nitro-host/src/nitro-host/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+authors = ["The Veracruz Development Team"]
+description = "A dummy package"
+edition = "2018"
+name = "nitro-host"
+version = "0.3.0"
+
+[dependencies]
+tests = { path = "../../crates/tests", features = [ "nitro" ] }
+veracruz-client = { path = "../../crates/veracruz-client", features = [ "nitro", "cli" ] }
+veracruz-server = { path = "../../crates/veracruz-server", features = [ "nitro", "cli" ] }
