+++ workspaces/nitro-runtime/Makefile
@@ -25,10 +25,10 @@ all: nitro
 nitro: runtime-manager-enclave
 	$(MAKE) runtime_manager.eif PCR0
 
-runtime_manager.eif PCR0 &: target/$(ARCH)-unknown-linux-musl/$(PROFILE_PATH)/runtime_manager_enclave crates/runtime-manager/dockerdir/Dockerfile
+runtime_manager.eif PCR0 &: target/$(ARCH)-unknown-linux-musl/$(PROFILE_PATH)/runtime-manager-enclave crates/runtime-manager/dockerdir/Dockerfile
 	rm -rf docker
 	mkdir -p docker
-	cp target/$(ARCH)-unknown-linux-musl/$(PROFILE_PATH)/runtime_manager_enclave docker
+	cp target/$(ARCH)-unknown-linux-musl/$(PROFILE_PATH)/runtime-manager-enclave docker
 	cp crates/runtime-manager/dockerdir/Dockerfile docker
 	perl -i -pe "s/x86_64/$(ARCH)/" docker/Dockerfile
 	nitro-cli build-enclave --docker-dir docker --docker-uri runtime_manager --output-file runtime_manager.eif > measurements.json
@@ -37,28 +37,17 @@ runtime_manager.eif PCR0 &: target/$(ARCH)-unknown-linux-musl/$(PROFILE_PATH)/ru
 
 runtime-manager-enclave:
 	rustup target add $(ARCH)-unknown-linux-musl
-	CC_$(ARCH)_unknown_linux_musl=$(ARCH)-linux-musl-gcc \
-	CXX_$(ARCH)_unknown_linux_musl=$(ARCH)-linux-musl-g++ \
-	RUSTFLAGS="-C link-args=-Wl,--build-id=none" \
-	cargo build --target $(ARCH)-unknown-linux-musl $(PROFILE_FLAG) $(V_FLAG) \
-		--features nitro -p runtime_manager_enclave
-	strip -p target/$(ARCH)-unknown-linux-musl/$(PROFILE_PATH)/runtime_manager_enclave
+	cargo build --target $(ARCH)-unknown-linux-musl $(PROFILE_FLAG) $(V_FLAG)
+	strip -p target/$(ARCH)-unknown-linux-musl/$(PROFILE_PATH)/runtime-manager-enclave
 
 doc:
 	cargo doc
 
 clippy:
 	rustup target add $(ARCH)-unknown-linux-musl
-	CC_$(ARCH)_unknown_linux_musl=$(ARCH)-linux-musl-gcc \
-	CXX_$(ARCH)_unknown_linux_musl=$(ARCH)-linux-musl-g++ \
 	cargo clippy --target $(ARCH)-unknown-linux-musl $(PROFILE_FLAG) $(V_FLAG) \
-		-p runtime_manager_enclave -p execution-engine \
+		-p nitro-runtime -p runtime-manager -p execution-engine \
 		-p session-manager -p policy-utils -p platform-services \
-		--features runtime_manager_enclave/nitro \
-		--features execution-engine/nitro  \
-		--features session-manager/nitro \
-		--features policy-utils/std \
-		--features platform-services/nitro \
 		-- --no-deps
 
 fmt:
